+++ midealocal/cli.py
@@ -15,10 +15,10 @@ import platformdirs
 from colorlog import ColoredFormatter
 
 from midealocal.cloud import SUPPORTED_CLOUDS, MideaCloud, get_midea_cloud
-from midealocal.device import MideaDevice, ProtocolVersion
+from midealocal.device import AuthException, MideaDevice, ProtocolVersion, RefreshFailed
 from midealocal.devices import device_selector
 from midealocal.discover import discover
-from midealocal.exceptions import ElementMissing
+from midealocal.exceptions import ElementMissing, SocketException
 from midealocal.version import __version__
 
 _LOGGER = logging.getLogger("cli")
@@ -65,13 +65,14 @@ class MideaCLI:
 
         return {**cloud_keys, **default_keys}
 
-    async def discover(self) -> MideaDevice | None:
+    async def discover(self) -> list[MideaDevice]:
         """Discover device information."""
         devices = discover(ip_address=self.namespace.host)
 
+        device_list: list[MideaDevice] = []
         if len(devices) == 0:
             _LOGGER.error("No devices found.")
-            return None
+            return device_list
 
         # Dump only basic device info from the base class
         _LOGGER.info("Found %d devices.", len(devices))
@@ -96,13 +97,24 @@ class MideaCLI:
                     subtype=0,
                     customize="",
                 )
-                _LOGGER.debug("Trying to connect with key: %s", key)
+                _LOGGER.debug("Opening socket for device.")
                 if dev.connect():
-                    _LOGGER.info("Found device:\n%s", dev.attributes)
-                    return dev
-
-                _LOGGER.debug("Unable to connect with key: %s", key)
-        return None
+                    try:
+                        if device["protocol"] == ProtocolVersion.V3:
+                            _LOGGER.debug("Trying to connect with key: %s", key)
+                            dev.authenticate()
+                        _LOGGER.debug("Trying to retrieve device attributes.")
+                        dev.refresh_status(True)
+                    except AuthException:
+                        _LOGGER.debug("Unable to connect with key: %s", key)
+                    except SocketException:
+                        _LOGGER.exception("Device socket closed.")
+                    except RefreshFailed:
+                        _LOGGER.exception("Unable to retrieve device attributes.")
+                    else:
+                        _LOGGER.info("Found device:\n%s", dev.attributes)
+                        device_list.append(dev)
+        return device_list
 
     def message(self) -> None:
         """Load message into device."""
@@ -168,23 +180,23 @@ class MideaCLI:
 
     async def set_attribute(self) -> None:
         """Set attribute for device."""
-        device = await self.discover()
-        if device is None:
+        device_list = await self.discover()
+        if len(device_list) != 1:
             return
 
         _LOGGER.info(
             "Setting attribute %s for %s [%s]",
             self.namespace.attribute,
-            device.device_id,
-            device.device_type,
+            device_list[0].device_id,
+            device_list[0].device_type,
         )
-        device.set_attribute(
+        device_list[0].set_attribute(
             self.namespace.attribute,
             self._cast_attr_value(),
         )
         await asyncio.sleep(2)
-        device.refresh_status(True)
-        _LOGGER.info("New device status:\n%s", device.attributes)
+        device_list[0].refresh_status(True)
+        _LOGGER.info("New device status:\n%s", device_list[0].attributes)
 
     def _cast_attr_value(self) -> int | bool | str:
         if self.namespace.attr_type == "bool":