+++ midealocal/devices/c3/__init__.py
@@ -2,7 +2,7 @@
 
 import json
 import logging
-from typing import Any
+from typing import Any, ClassVar
 
 from midealocal.device import MideaDevice
 from midealocal.devices.c3.const import C3DeviceMode, C3SilentLevel, DeviceAttributes
@@ -22,6 +22,12 @@ _LOGGER = logging.getLogger(__name__)
 
 class MideaC3Device(MideaDevice):
     """Midea C3 device."""
+
+    _silent_modes: ClassVar[list[str]] = [
+        C3SilentLevel.OFF.name,
+        C3SilentLevel.SILENT.name,
+        C3SilentLevel.SUPER_SILENT.name,
+    ]
 
     def __init__(
         self,
@@ -62,7 +68,7 @@ class MideaC3Device(MideaDevice):
                 DeviceAttributes.zone1_water_temp_mode: False,
                 DeviceAttributes.zone2_water_temp_mode: False,
                 DeviceAttributes.silent_mode: False,
-                DeviceAttributes.SILENT_LEVEL: C3SilentLevel.OFF,
+                DeviceAttributes.SILENT_LEVEL: C3SilentLevel.OFF.name,
                 DeviceAttributes.eco_mode: False,
                 DeviceAttributes.tbh: False,
                 DeviceAttributes.mode: 1,
@@ -100,6 +106,11 @@ class MideaC3Device(MideaDevice):
     def temperature_step(self) -> float | None:
         """Midea C3 device temperature step."""
         return self._temperature_step
+
+    @property
+    def silent_modes(self) -> list[str]:
+        """Midea C3 device silent modes."""
+        return MideaC3Device._silent_modes
 
     def build_query(self) -> list[MessageQuery]:
         """Midea C3 device build query."""
@@ -241,11 +252,15 @@ class MideaC3Device(MideaDevice):
             if attr == DeviceAttributes.silent_mode.value and isinstance(value, bool):
                 message.silent_mode = bool(value)
                 message.silent_level = (
-                    C3SilentLevel.SILENT if value else C3SilentLevel.OFF
+                    C3SilentLevel.SILENT
+                    if value
+                    and self._attributes[DeviceAttributes.SILENT_LEVEL]
+                    == C3SilentLevel.OFF.name
+                    else C3SilentLevel[self._attributes[DeviceAttributes.SILENT_LEVEL]]
                 )
-            elif attr == DeviceAttributes.SILENT_LEVEL.value and isinstance(value, int):
-                message.silent_level = C3SilentLevel(int(value))
-                message.silent_mode = value != C3SilentLevel.OFF
+            elif attr == DeviceAttributes.SILENT_LEVEL.value and isinstance(value, str):
+                message.silent_level = C3SilentLevel[value]
+                message.silent_mode = value != C3SilentLevel.OFF.name
         if message is not None:
             self.build_send(message)