+++ midealocal/devices/c3/message.py
@@ -240,12 +240,12 @@ class C3QuerySilenceMessageBody(MessageBody):
     def __init__(self, body: bytearray, data_offset: int = 0) -> None:
         """Initialize C3 notify1 message body."""
         super().__init__(body)
-        self.silence_mode = body[data_offset] & 0x1 > 0
-        self.silence_level = (
+        self.silent_mode = body[data_offset] & 0x1 > 0
+        self.silent_level = C3SilentLevel(
             (body[data_offset] & 0x1) + ((body[data_offset] & 0x8) >> 2)
-            if self.silence_mode
-            else C3SilentLevel.OFF
-        )
+            if self.silent_mode
+            else C3SilentLevel.OFF.value,
+        ).name
         # Message protocol information:
         # silence_function_state: Byte 1, BIT 0
         # silence_timer1_state: Byte 1, BIT 1