+++ tests/devices/c3/device_c3_test.py
@@ -51,7 +51,8 @@ class TestMideaC3Device:
         assert self.device.attributes[DeviceAttributes.zone2_water_temp_mode] is False
         assert self.device.attributes[DeviceAttributes.silent_mode] is False
         assert (
-            self.device.attributes[DeviceAttributes.SILENT_LEVEL] == C3SilentLevel.OFF
+            self.device.attributes[DeviceAttributes.SILENT_LEVEL]
+            == C3SilentLevel.OFF.name
         )
         assert self.device.attributes[DeviceAttributes.eco_mode] is False
         assert self.device.attributes[DeviceAttributes.tbh] is False
@@ -90,6 +91,7 @@ class TestMideaC3Device:
         assert self.device.attributes[DeviceAttributes.outdoor_temperature] is None
         assert self.device.attributes[DeviceAttributes.error_code] == 0
         assert self.device.temperature_step == 1
+        assert len(self.device.silent_modes) == 3
 
     def test_set_attribute(self) -> None:
         """Test set attribute."""
@@ -108,7 +110,7 @@ class TestMideaC3Device:
 
             self.device.set_attribute(
                 DeviceAttributes.SILENT_LEVEL.value,
-                C3SilentLevel.SILENT,
+                C3SilentLevel.SILENT.name,
             )
 
     def test_build_query(self) -> None: